+++ tests/test/Main.fs
@@ -1,0 +1,14 @@
+module Main
+
+
+#if !HEADLESS
+
+HelloWorldTest.init()
+DOMTest.init()
+BindingTest.init()
+ObservableTest.init()
+StoreTest.init()
+
+BrowserFramework.runAll()
+
+#endif